requirements/test.txt
-r base.txt

pytest-django==4.11.1
pytest-cov==7.0.0
